+++ src/services/notifications/notification/index.ts
@@ -1,0 +1,7 @@
+// src/services/notifications/notification/index.ts
+
+export {
+  notificationService,
+  NotificationServiceClass,
+} from "./NotificationService";
+export { NotificationRepository } from "./NotificationRepository";